Implement news fetching and classification for major tech companies
- Added functionality to fetch news for a list of companies over the past 7 years.
- Introduced a classification method to categorize news articles by sentiment.
- Integrated logging to track the progress and performance of the news processing.
- Visualized sentiment results for each company and aggregated totals.
- Improved performance measurement for news classification.
