+++ app/modelo_fundamental/__init__.py
@@ -1,0 +1,2 @@
+from .cliente_clasificador import Empresa, empresas_lit, graficas
+from .get_noticias import FetchNews
