--- app/modelo_lenguaje/README.md
@@ -1,10 +1,0 @@
-# modelo_lenguaje
-
-Este módulo implementa el análisis fundamental del sistema. Se encarga de procesar información textual relevante para el mercado financiero.
-
-## Subcarpetas
-
-- `analizador/`: Procesa y evalúa textos (noticias, reportes, comunicados) para asignarles una puntuación de confianza, alerta o riesgo.
-- `extractor/`: Extrae información relevante de textos para alimentar el análisis fundamental.
-
-El objetivo es transformar la narrativa del mercado en señales cuantificables que puedan integrarse con el análisis técnico.
