--- app/modelo_lenguaje/analizador/README.md
@@ -1,10 +1,0 @@
-# analizador
-
-Este submódulo del análisis fundamental se encarga de procesar y evaluar textos extraídos de fuentes relevantes (noticias, reportes, comunicados).
-
-## Propósito
-
-- Analizar el contenido textual para identificar el sentimiento (confianza, alerta, riesgo).
-- Asignar puntuaciones a cada noticia o reporte según su impacto potencial en el mercado.
-
-El resultado es una valoración cuantitativa de la narrativa del mercado.
