--- app/modelo_lenguaje/extractor/README.md
@@ -1,10 +1,0 @@
-# extractor
-
-Este submódulo del análisis fundamental se encarga de la extracción de información relevante de textos.
-
-## Propósito
-
-- Recolectar datos de noticias, comunicados, reportes y cualquier texto que hable sobre la empresa o el mercado.
-- Preparar la información para su posterior análisis y evaluación.
-
-El objetivo es alimentar el sistema con datos textuales actualizados y relevantes.
